Apply things from fix

+++ src/cst816s.rs
@@ -114,8 +114,8 @@ where
         touch.y = (buf[Self::TOUCH_Y_L_OFF] as i32) | (((touch_y_h_and_finger & 0x0F) as i32) << 8);
 
         // action of touch (0 = down, 1 = up, 2 = contact)
-        touch.action = (touch_x_h_and_action >> 6) as u8;
-        touch.finger_id = (touch_y_h_and_finger >> 4) as u8;
+        touch.action = touch_x_h_and_action >> 6;
+        touch.finger_id = touch_y_h_and_finger >> 4;
 
         //  Compute touch pressure and area
         touch.pressure = buf[Self::TOUCH_PRESURE_OFF];
@@ -128,8 +128,8 @@ where
     /// Returns a touch event if available.
     ///
     /// - `check_int_pin` -- True if we should check the interrupt pin before attempting i2c read.
-    /// On some devices, attempting to read registers when there is no data available results
-    /// in a hang in the i2c read.
+    ///   On some devices, attempting to read registers when there is no data available results
+    ///   in a hang in the i2c read.
     ///
     pub fn read_one_touch_event(&mut self, check_int_pin: bool) -> Option<TouchEvent> {
         let mut one_event: Option<TouchEvent> = None;
